+++ source/common/network/udp_packet_writer_handler_impl.cc
@@ -6,8 +6,6 @@
 namespace Envoy {
 namespace Network {
 
-REGISTER_FACTORY(UdpDefaultWriterFactoryFactory, UdpPacketWriterFactoryFactory);
-
 UdpDefaultWriter::UdpDefaultWriter(Network::IoHandle& io_handle)
     : write_blocked_(false), io_handle_(io_handle) {}
