+++ source/common/network/udp_packet_writer_handler_impl.h
@@ -1,10 +1,8 @@
 #pragma once
 
 #include "envoy/buffer/buffer.h"
-#include "envoy/extensions/udp_packet_writer/v3/udp_default_writer_factory.pb.h"
 #include "envoy/network/socket.h"
 #include "envoy/network/udp_packet_writer_handler.h"
-#include "envoy/registry/registry.h"
 
 #include "source/common/network/io_socket_error_impl.h"
 
@@ -50,20 +48,6 @@ public:
     return std::make_unique<UdpDefaultWriter>(io_handle);
   }
 };
-
-class UdpDefaultWriterFactoryFactory : public Network::UdpPacketWriterFactoryFactory {
-public:
-  std::string name() const override { return "envoy.udp_packet_writer.default"; }
-  UdpPacketWriterFactoryPtr
-  createUdpPacketWriterFactory(const envoy::config::core::v3::TypedExtensionConfig&) override {
-    return std::make_unique<UdpDefaultWriterFactory>();
-  }
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override {
-    return std::make_unique<envoy::extensions::udp_packet_writer::v3::UdpDefaultWriterFactory>();
-  }
-};
-
-DECLARE_FACTORY(UdpDefaultWriterFactory);
 
 } // namespace Network
 } // namespace Envoy
