+++ source/common/quic/udp_gso_batch_writer.cc
@@ -112,7 +112,5 @@ UdpGsoBatchWriterFactory::createUdpPacketWriter(Network::IoHandle& io_handle, St
   return std::make_unique<UdpGsoBatchWriter>(io_handle, scope);
 }
 
-REGISTER_FACTORY(UdpGsoBatchWriterFactoryFactory, Network::UdpPacketWriterFactoryFactory);
-
 } // namespace Quic
 } // namespace Envoy
